# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sinema API",
        "description": "Film, tür, kişi ve oyuncu yönetimi için RESTful API",
        "contact": {
            "name": "API Geliştirici",
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/kisiler": {
            "get": {
                "tags": [
                    "Kişiler"
                ],
                "summary": "Tüm kişileri listele",
                "description": "Veritabanındaki tüm kişileri (oyuncu, yönetmen vs.) getirir",
                "operationId": "7ae244487d3e2b0d2f8bee45a97f1052",
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Kisi"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Kişiler"
                ],
                "summary": "Yeni kişi ekle",
                "description": "Veritabanına yeni bir kişi (oyuncu, yönetmen vs.) ekler",
                "operationId": "35f01d8a4fcff50fa35a926bc8ef46e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "adi"
                                ],
                                "properties": {
                                    "adi": {
                                        "description": "Kişi adı",
                                        "type": "string",
                                        "example": "Keanu Reeves"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kişi başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Kisi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                }
            }
        },
        "/kisiler/{id}": {
            "get": {
                "tags": [
                    "Kişiler"
                ],
                "summary": "Belirli kişiyi getir",
                "description": "ID'ye göre belirli bir kişiyi getirir",
                "operationId": "9ca8b69a2aafb77d365d1609e2ccc888",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kişi ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Kisi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kişi bulunamadı"
                    }
                }
            },
            "put": {
                "tags": [
                    "Kişiler"
                ],
                "summary": "Kişi güncelle",
                "description": "Belirli bir kişinin bilgilerini günceller",
                "operationId": "8cef6ce8c043bb8e4e422e98902bdc83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kişi ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "adi"
                                ],
                                "properties": {
                                    "adi": {
                                        "description": "Kişi adı",
                                        "type": "string",
                                        "example": "Keanu Reeves"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kişi başarıyla güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Kisi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kişi bulunamadı"
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Kişiler"
                ],
                "summary": "Kişi sil",
                "description": "Belirli bir kişiyi veritabanından siler",
                "operationId": "259bdb940bd9ad86e8d876b596f01c37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kişi ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kişi başarıyla silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kişi başarıyla silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kişi bulunamadı"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Kisi": {
                "title": "Kişi",
                "description": "Kişi modeli (oyuncu, yönetmen vs.)",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "adi": {
                        "description": "Kişi adı",
                        "type": "string",
                        "example": "Keanu Reeves"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}